crypto-arbitrage-bot/requirements.txt
# Core
websockets>=12.0
aiohttp>=3.9.0
asyncio

# Exchanges
python-binance>=1.0.17
# coinbase>=2.1.0  # Optional, can use REST/WS directly
# krakenex>=2.1.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Monitoring
prometheus-client>=0.17.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
